Move is_json to json_handling.py
